requirements.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==2.4.1
asyncpg==0.30.0
bcrypt==4.2.0
billiard==4.2.1
celery==5.5.2
certifi==2024.12.14
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
executing==2.0.1
fastapi==0.115.6
fastapi-cache2==0.2.2
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.26.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.5.3
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
nodeenv==1.9.1
orjson==3.10.13
packaging==25.0
parso==0.8.4
passlib==1.7.4
pendulum==3.1.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.47
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyright==1.1.401
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.1.0
redis==6.1.0
rich==13.9.4
rich-toolkit==0.12.0
ruff==0.11.11
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
